add support to multiple exchanges (tangled and fiatleak exchange)

+++ src/api/QBXWsEGSGthDhsLn/index.js
@@ -1,6 +1,6 @@
 import Endpoint from '../endpoint';
 import database from '../../database/database';
-import TangledExchangeApi from '../tangled-exchange-api';
+import ExchangeApi from '../exchange-api';
 
 
 /**
@@ -18,10 +18,17 @@ class _QBXWsEGSGthDhsLn extends Endpoint {
      * @param res
      */
     handler(app, req, res) {
+        const exchange = this.exchangeIds[p0?.toLowerCase()];
+        if (!exchange) {
+            return res.status(400).send({
+                api_status : 'fail',
+                api_message: 'p0<exchange>[tangled.com or fiatleak.com] is required'
+            });
+        }
         const configRepository = database.getRepository('config');
         configRepository.deleteConfig('tangled_exchange_api_key')
                         .then(() => {
-                            TangledExchangeApi.setApiKey(undefined);
+                            ExchangeApi.get(exchange).setApiKey(undefined);
                             res.send({
                                 api_status: 'success'
                             });
